requirements.txt
flask==2.2.3
flask-login==0.6.2
flask-wtf==1.1.1
flask-pymongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
werkzeug==2.2.3
flask-misaka==1.0.0
markdown==3.4.3
pygments==2.15.1
flask-mail==0.9.1
flask-caching==2.0.2
pillow==9.5.0
numpy==1.24.3
beautifulsoup4==4.12.2
email-validator==2.0.0
bcrypt==4.0.1
passlib==1.7.4
pytz==2023.3
gunicorn==20.1.0 